requirements.txt
django==4.2.4
djangorestframework==3.14
django-debug-toolbar==3.2.4
